테스트 추가

+++ Assignment3/Utils.h
@@ -7,6 +7,6 @@ namespace assignment3
 	template<typename T>
 	T round(T number)
 	{
-		return static_cast<T>(std::round((double)number * 1000.0) / 1000.0);
+		return static_cast<T>(std::round(number * 1000) / 1000);
 	}
 }
